Restrcuture the student dashboard
