+++ components/dashboard/studentNav.js
@@ -6,55 +6,55 @@ import Axios from "axios";
 import { useRouter } from "next/router";
 
 function studentNav() {
-  const router = useRouter();
+	const router = useRouter();
 
-  function logOut() {
-    Axios.post("/api/logout")
-      .then(() => {
-        console.log("success");
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }
-  return (
-    <nav className={navbarStyles.navBar}>
-      <img className={navbarStyles.boomerangImg} src={"/smallLogo.png"} />
+	function logOut() {
+		Axios.post("/api/logout")
+			.then(() => {
+				console.log("success");
+			})
+			.catch((err) => {
+				console.log(err);
+			});
+	}
+	return (
+		<nav className={navbarStyles.navBar}>
+			<img
+				className={navbarStyles.boomerangImg}
+				src={"/smallLogo.png"}
+			/>
 
-      <span className={navbarStyles.menu}>
-        <Link href="/app/dashboard/student">
-          <a className={navbarStyles.menuItem} title="participants">
-            Home
-          </a>
-        </Link>
-        <Link href="/app/dashboard/student/tasklist">
-          <a className={navbarStyles.menuItem} title="participants">
-            My Tasks
-          </a>
-        </Link>
-        <Link href="/app/dashboard/student/calendar">
-          <a className={navbarStyles.menuItem} title="Data">
-            My Calendar
-          </a>
-        </Link>
-        <Link href="/app/dashboard/student/questionnaire">
-          <a className={navbarStyles.menuItem} title="participants">
-            My Questionnare
-          </a>
-        </Link>
-        <Link href="/app/login">
-          <a onClick={logOut} className={navbarStyles.menuItem} title="logout">
-            Log out
-          </a>
-        </Link>
-        {/*
+			<span className={navbarStyles.menu}>
+				<Link href='/app/dashboard/student'>
+					<a
+						className={navbarStyles.menuItem}
+						title='participants'>
+						Home
+					</a>
+				</Link>
+				<Link href='/app/dashboard/student/questionnaire'>
+					<a
+						className={navbarStyles.menuItem}
+						title='participants'>
+						My Questionnare
+					</a>
+				</Link>
+				<Link href='/app/login'>
+					<a
+						onClick={logOut}
+						className={navbarStyles.menuItem}
+						title='logout'>
+						Log out
+					</a>
+				</Link>
+				{/*
                 <Link href='/app/dashboard/student/questionnare'>
                     <a title='Referral'>Questionnare</a>
                 </Link>
                 */}
-      </span>
-    </nav>
-  );
+			</span>
+		</nav>
+	);
 }
 
 export default studentNav;
